Protect RFKILL under CONFIG_RFKILL
Hi,

We are working on a platform which will not support RFKILL functionality.
So we tried to disable RFKILL in wpa_supplicant. Disabling NEED_RFKILL
config option in drivers.mak and drivers.mk was not disabling the
complete RFKILL functionality in wpa_supplicant. So provided an option
(CONFIG_RFKILL) in defconfig to enable and disable rfkill functionality in
wpa_supplicant. CONFIG_RFKILL macro is used to protect RFKILL part of code
in wpa_supplicant.

RFKILL functionality disable option is required in the case
where execution platform doesnot support RFKILL functionality.

Signed-off-by: Harsha Sakre P <ps.harsha@globaledgesoft.com>
Message-Id: <1491469809-11620-1-git-send-email-ps.harsha@globaledgesoft.com>