Make ICV Indicator dependant on ICV length
IEEE 802.1X-2010 Section 11.11 describes that the ICV is seperate from
the parameter sets before it. Due to its convenient layout the ICV
Indicator 'body part' is used to encode the ICV as well.

IEEE 802.1X-2010 Section 11.11.3 describes the encoding of MKPDUs.
In bullet e) is desribed that the ICV Indicator itself is encoded when
the ICV is not 16 octets in length. IEEE 802.1Xbx-2014 Table 11.7 note
e) states that it will not be encoded unless the Algorithm Agility
parameter specifies the use of an ICV that is not 16 octets in length.

Therefore the length calculation for the ICV indicator body part must
take into account if the ICV Indicator is to be encoded or not. The
actual encoder of the ICV body already takes care of the rest.

Signed-off-by: Jaap Keuter <jaap.keuter@xs4all.nl>
Message-Id: <20170407213923.30891-1-jaap.keuter@xs4all.nl>