nl80211: Fix race condition in detecting MAC change
Commit 3e0272ca00ce ("nl80211: Re-read MAC address on RTM_NEWLINK")
added the detection of external changes to MAC address when the
interface is brought up.

If the interface state is changed quickly enough, wpa_supplicant may
receive the netlink message for the !IFF_UP event when the interface
has already been brought up and would ignore the next netlink IFF_UP
message, missing the MAC change.

Fix this by also reloading the MAC address when a !IFF_UP event is
received with the interface up, because this implies that the
interface went down and up again, possibly changing the address.

Signed-off-by: Beniamino Galvani <bgalvani@redhat.com>
Message-Id: <1493395170-13494-1-git-send-email-bgalvani@redhat.com>