crypto: add option to use getrandom()
According to random(4) manual, /dev/random is essentially deprecated on Linux
for quite some time:

  The /dev/random interface is considered a legacy interface, and
  /dev/urandom is preferred and sufficient in all use cases, with the
  exception of applications which require randomness during early boot time;
  for these applications, getrandom(2) must be used instead, because it will
  block until the entropy pool is initialized.

An attempt to use it would cause unnecessary blocking on machines
without a good hwrng even when it shouldn't be needed. Since Linux 3.17,
a getrandom(2) call is available that will block only until the
randomness pool has been seeded.

It is probably not a good default yet as it requires a fairly recent
kernel and glibc (3.17 and 2.25 respectively).

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>
Message-Id: <20170918125807.27940-1-lkundrak@v3.sk>