wpa_supplicant: update MAC when driver detects a change
wpa_supplicant updates own_addr when the interface becomes enabled, as
the MAC can only change when the interface is down. However, drivers
don't report all interface state changes: for example the nl80211
driver may ignore a down-up cycle if the down message is processed
when the interface is already up.

In such cases, wpa_supplicant (and in particular, the EAP state
machine) would continue to use the old MAC causing authentication
failures due to a mismatch of the computed PMKID.

Add a new driver event that notifies of MAC address changes while the
interface is active.

Signed-off-by: Beniamino Galvani <bgalvani@redhat.com>
Message-Id: <20171030094313.30272-1-bgalvani@redhat.com>