dbus: export roam time, roam complete, and session length to dbus
Added new Interface properties "RoamTime", "RoamComplete", and "SessionLength".

"RoamTime" carries the roam time of the most recent roam in milliseconds.
"RoamComplete" carries True or False corresponding to the success status of the
most recent roam.
"SessionLength" carries the number of milliseconds corresponding to how long
the connection to the last AP was before a roam or disconnect happened.

These properties will all be logged in UMA to compare FT vs non-FT network
behaviors.

Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Message-Id: <20180504181602.34764-1-matthewmwang@chromium.org>