openssl: load 'legacy' provider with openssl-3.0.0
openssl-3.0.0 deprecated some algorithms that are still widely used by
wpa_supplicant. For example, MSCHAPv2 specifies use of MD4 / RC4 / DES,
and all them are no more available in the "default" provider.

load the "legacy" provider when initializing the openssl backend, in
case detected version is greater or equal to 3.0.0

Signed-off-by: Davide Caratti <davide.caratti@gmail.com>